backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
reportlab
python-multipart
python-jose[cryptography]
gunicorn
psycopg2-binary
